// src/order/types.ts
export enum OrderStatus {
  PENDING_CONFIRMATION = 'PENDING_CONFIRMATION',
  REJECTED = 'REJECTED',
  CONFIRMED = 'CONFIRMED',
  SHIPPED = 'SHIPPED',
  DELIVERED = 'DELIVERED',
  CANCELLED = 'CANCELLED',
  CANCELLED_BY_DRIVER = 'CANCELLED_BY_DRIVER',
  CANCELLED_BY_CLIENT = 'CANCELLED_BY_CLIENT',
  CANCELLED_BY_ADMIN = 'CANCELLED_BY_ADMIN',
}

export enum OrderStatusTransition {
  CONFIRM = 'CONFIRM',
  REJECT = 'REJECT',
  SHIP = 'SHIP',
  DELIVER = 'DELIVER',
  CANCEL_BY_ADMIN = 'CANCEL_BY_ADMIN',
  CANCEL_BY_DRIVER = 'CANCEL_BY_DRIVER',
  CANCEL_BY_CLIENT = 'CANCEL_BY_CLIENT',
}